Move all scripts under RSL folder
