+++ Assets/RSL/Core/Interaction/Scripts.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b76564ffa0f050c1eb2a9a4f27e0e70b
+guid: 83831a12dceeb8d40b18ade886b78963
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
